fix: update B3trBalance and related services to use BigDecimal for ba… (#1200)

+++ packages/common/src/main/kotlin/org/vechain/indexer/b3tr/balance/B3trBalance.kt
@@ -3,7 +3,7 @@ package org.vechain.indexer.b3tr.balance
 import com.fasterxml.jackson.annotation.JsonIgnore
 import com.fasterxml.jackson.annotation.JsonInclude
 import com.fasterxml.jackson.annotation.JsonView
-import java.math.BigInteger
+import java.math.BigDecimal
 import org.springframework.data.annotation.Id
 import org.springframework.data.mongodb.core.mapping.Document
 import org.springframework.data.mongodb.core.mapping.Field
@@ -21,9 +21,9 @@ data class B3trBalance(
     @JsonIgnore override val blockNumber: Long,
     @JsonIgnore override val blockTimestamp: Long,
     @JsonIgnore @field:JsonView(Views.Internal::class) override val version: Int,
-    @Field(targetType = FieldType.DECIMAL128) var vot3Balance: BigInteger,
-    @Field(targetType = FieldType.DECIMAL128) var b3trBalance: BigInteger,
-    @Field(targetType = FieldType.DECIMAL128) var totalBalance: BigInteger,
+    @Field(targetType = FieldType.DECIMAL128) var vot3Balance: BigDecimal,
+    @Field(targetType = FieldType.DECIMAL128) var b3trBalance: BigDecimal,
+    @Field(targetType = FieldType.DECIMAL128) var totalBalance: BigDecimal,
 ) : VersionedDocument {
     @JsonIgnore override fun getDocumentId(): String = address
 }
